refactor: move supplier module unit test from locations to suppliers app
